+++ packages/dashmate/configs/getConfigFileMigrationsFactory.js
@@ -2,17 +2,17 @@
 
 /**
  * @param {HomeDir} homeDir
- * @param {SystemConfigs} systemConfigs
+ * @param {DefaultConfigs} defaultConfigs
  * @returns {getConfigFileMigrations}
  */
-function getConfigFileMigrationsFactory(homeDir, systemConfigs) {
+function getConfigFileMigrationsFactory(homeDir, defaultConfigs) {
   /**
    * @typedef {function} getConfigFileMigrations
    * @returns {Object}
    */
   function getConfigFileMigrations() {
-    const base = systemConfigs.get('base');
-    const testnet = systemConfigs.get('testnet');
+    const base = defaultConfigs.get('base');
+    const testnet = defaultConfigs.get('testnet');
 
     return {
       '0.24.0': (configFile) => {
@@ -96,6 +96,21 @@ function getConfigFileMigrationsFactory(homeDir, systemConfigs) {
             }
           });
 
+        return configFile;
+      },
+      '0.24.17': (configFile) => {
+        Object.entries(configFile.configs)
+          .forEach(([, options]) => {
+            options.docker.baseImage = base.get('docker.baseImage');
+          });
+
+        return configFile;
+      },
+      '0.24.20': (configFile) => {
+        Object.entries(configFile.configs)
+          .forEach(([, options]) => {
+            options.core.docker.image = base.get('core.docker.image');
+          });
         return configFile;
       },
     };